feat: add authentication endpoints

+++ src/main/java/task_management_system/task/service/TaskService.java
@@ -5,14 +5,21 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import task_management_system.exception.BadRequestException;
+import task_management_system.exception.ForbiddenException;
 import task_management_system.exception.NotFoundException;
 import task_management_system.task.dto.CreateTaskRequest;
 import task_management_system.dto.CustomResponse;
 import task_management_system.task.dto.TaskDto;
 import task_management_system.task.dto.UpdateTask;
 import task_management_system.task.entity.Task;
+import task_management_system.task.entity.TaskRole;
+import task_management_system.task.enums.RoleType;
 import task_management_system.task.repository.TaskRepository;
+import task_management_system.task.repository.TaskRoleRepository;
+import task_management_system.user.entity.User;
+import task_management_system.utils.Utils;
 
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
@@ -24,12 +31,19 @@ import java.util.UUID;
 public class TaskService {
 
     private final TaskRepository taskRepository;
+    private final TaskRoleRepository taskRoleRepository;
     private static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss");
 
+    @Transactional
     public TaskDto createTask(CreateTaskRequest taskRequest) {
+        User authUser = Utils.getAuthenticatedUser();
         LocalDateTime dueDate;
+
         try {
             dueDate = LocalDateTime.parse(taskRequest.getDueDate(), DATE_TIME_FORMATTER);
+            if (dueDate.isBefore(LocalDateTime.now())) {
+                throw new BadRequestException("due date must be in the future to be valid");
+            }
         } catch (DateTimeException dte) {
             throw new BadRequestException(dte.getMessage());
         }
@@ -41,6 +55,7 @@ public class TaskService {
                 .status(taskRequest.getStatus())
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
+                .createdBy(authUser)
                 .priority(taskRequest.getPriority())
                 .assignedTo(taskRequest.getAssignedTo())
                 .tags(taskRequest.getTags())
@@ -51,19 +66,41 @@ public class TaskService {
     }
 
     public TaskDto getTaskByID(UUID taskID) {
-        Task task = findTaskByID(taskID);
+        User authUser = Utils.getAuthenticatedUser();
+        Task task = findTaskByID(taskID, authUser);
+
         return convertToDo(task);
     }
 
     public Page<TaskDto> getTasks(int page, int size) {
+        User authUser = Utils.getAuthenticatedUser();
+
         Pageable pageable = PageRequest.of(page, size);
-        Page<Task> tasks = taskRepository.findAll(pageable);
+        Page<Task> tasks = taskRepository.findTasksByUserRoles(
+                authUser.getId(),
+                pageable
+        );
 
         return tasks.map(this::convertToDo);
     }
 
     public CustomResponse updateTask(UUID taskID, UpdateTask updateRequest) {
-        Task task = findTaskByID(taskID);
+        User authUser = Utils.getAuthenticatedUser();
+        Task task = findTaskByID(taskID, authUser);
+
+        TaskRole role = taskRoleRepository.findByTaskAndUser(task, authUser)
+                .orElseThrow(() -> new ForbiddenException("user has no role on this task"));
+
+        if (!role.getRoleType().equals(RoleType.CREATOR)) {
+            if (updateRequest.getTitle() != null ||
+                    updateRequest.getDescription() != null ||
+                    updateRequest.getDueDate() != null ||
+                    updateRequest.getPriority() != null ||
+                    updateRequest.getAssignedTo() != null ||
+                    updateRequest.getTags() != null) {
+                throw new ForbiddenException("Only the status can be updated by assigned users.");
+            }
+        }
 
         task.setTitle(updateRequest.getTitle() != null
                 ? updateRequest.getTitle() : task.getTitle());
@@ -90,7 +127,16 @@ public class TaskService {
     }
 
     public CustomResponse deleteTask(UUID taskID) {
-        Task task = findTaskByID(taskID);
+        User authUser = Utils.getAuthenticatedUser();
+        Task task = findTaskByID(taskID, authUser);
+
+        TaskRole role = taskRoleRepository.findByTaskAndUser(task, authUser)
+                .orElseThrow();
+
+        boolean hasAuthority = role.getRoleType().equals(RoleType.CREATOR);
+        if (!hasAuthority) {
+            throw new ForbiddenException("only creator of task can delete task");
+        }
 
         taskRepository.delete(task);
         return CustomResponse.builder()
@@ -99,13 +145,24 @@ public class TaskService {
                 .build();
     }
 
-    private Task findTaskByID(UUID taskID) {
-        return taskRepository.findById(taskID)
+    private Task findTaskByID(UUID taskID, User authUser) {
+
+        Task task = taskRepository.findById(taskID)
                 .orElseThrow(() -> new NotFoundException("Task not found with id: " +  taskID));
+
+        boolean hasRole = taskRoleRepository.existsByTaskIdAndUserId(
+                taskID, authUser.getId()
+        );
+
+        if (!hasRole) {
+            throw new ForbiddenException("You are not authorized to access this resource");
+        }
+
+        return task;
     }
 
     private TaskDto convertToDo(Task task) {
-        return  TaskDto.builder()
+        return TaskDto.builder()
                 .id(task.getId())
                 .title(task.getTitle())
                 .description(task.getDescription())
@@ -114,7 +171,7 @@ public class TaskService {
                 .dueDate(task.getDueDate())
                 .tags(task.getTags())
                 .priority(task.getPriority())
-                .createdBy(null)
+                .createdBy(task.getCreatedBy().getId())
                 .createdAt(task.getCreatedAt())
                 .updatedAt(task.getUpdatedAt())
                 .build();
